+++ lib/helpers/json_helper.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 
+import 'package:money/helpers/misc_helpers.dart';
+
 typedef MyJson = Map<String, dynamic>;
 
 /// helpers for retrieving value as types
@@ -27,7 +29,10 @@ extension MyJsonExtensions on MyJson {
       return defaultIfNotFound;
     }
     try {
-      return value as int;
+      if (value is int) {
+        return value;
+      }
+      return int.parse(value);
     } catch (_) {
       return defaultIfNotFound;
     }
@@ -45,6 +50,9 @@ extension MyJsonExtensions on MyJson {
       if (value is int) {
         return (value).toDouble();
       }
+      if (value is String) {
+        return attemptToGetDoubleFromText(value) ?? 0.0;
+      }
       return value as double;
     } catch (_) {
       return defaultIfNotFound;
@@ -60,6 +68,10 @@ extension MyJsonExtensions on MyJson {
       return defaultIfNotFound;
     }
     try {
+      if (value is bool) {
+        return value;
+      }
+
       return value as bool;
     } catch (_) {
       return defaultIfNotFound;
@@ -109,4 +121,12 @@ String encodeValueWrapStringTypes(dynamic value) {
   } else {
     return value.toString();
   }
+}
+
+MyJson myJsonFromKeyValuePairs(List<String> keys, List<String> values) {
+  MyJson object = {};
+  for (int i = 0; i < keys.length; i++) {
+    object[keys[i]] = values[i];
+  }
+  return object;
 }